requirements.txt
Flask
requests
Flask-Cors
Pillow
pandas
folium
line-bot-sdk